A short shell script I wrote to start up the Tor Browser Bundle for Linux in a slightly unusual way.
I routinely install Tor from my distro's native package repositories, so I
don't use the copy that comes with the TBB.  However, starting Firefox then
becomes something of a problem.  I solved this by figuring out how to start
Firefox with the correct user profile (which will make use of whatever
instance of Tor it finds on port 9001/tcp).

Short, sweet, and to the point.  If you don't want to do it this way, you
don't need this script, use the one that comes with the TBB.

Signed-off-by: The Doctor <drwho@virtadpt.net>